+++ gulpfile.js
@@ -67,6 +67,7 @@ gulp.task('js-merge', function () {
     './src/74update.js',
     './src/74update.js',
     './src/76usedatabase.js',
+    './src/77declare.js',
     './src/78show.js',
     './src/79set.js',
     './src/80console.js',